PureLB and dnsmasq
dnsmasq is installed on extra IP that is registered in Fritzbox and
used as DNS server in the home network.

This allows to add own host names e.g. for  home assistant or k3s.fritz.box
that resolves to all k3s nodes.

dnsmasq is setup redundant on each node. PureLB ensures that failover.
Required to not loose DNS when updating the OS of the nodes.